Split golf panel UI
