+++ src/plugins/code-golf/GolfStatsPanel.less
